+++ jax/_src/scipy/linalg.py
@@ -875,22 +875,6 @@ def polar(a: ArrayLike, side: str = 'right', *, method: str = 'qdwh', eps: Optio
   return unitary, posdef
 
 
-def polar_unitary(a: ArrayLike, *, method: str = "qdwh", eps: Optional[float] = None,
-                  max_iterations: Optional[int] = None) -> Tuple[Array, Array]:
-  """ Computes the unitary factor u in the polar decomposition ``a = u p``
-  (or ``a = p u``).
-
-  .. warning::
-    This function is deprecated. Use :func:`jax.scipy.linalg.polar` instead.
-  """
-  # TODO(phawkins): delete this function after 2022/8/11.
-  warnings.warn("jax.scipy.linalg.polar_unitary is deprecated. Call "
-                "jax.scipy.linalg.polar instead.",
-                DeprecationWarning)
-  unitary, _ = polar(a, method, eps, max_iterations)
-  return unitary
-
-
 @jit
 def _sqrtm_triu(T: Array) -> Array:
   """